Handled exceptions and returning correct custom messages in error body

+++ src/main/java/com/gloal/todo/service/TodoServiceImpl.java
@@ -13,7 +13,6 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 @Service
 public class TodoServiceImpl implements TodoService {
@@ -23,48 +22,47 @@ public class TodoServiceImpl implements TodoService {
 
     @Override
     public List<Todo> fetchAllTodos() {
+        //get all todos or return empty list if server error
         try {
             return todoRepository.findAll();
-        }catch (Exception e){
-           LOGGER.error("Error retrieving todo list", e);
+        } catch (Exception e) {
+            LOGGER.error("Error retrieving todo list", e);
         }
         return new ArrayList<>();
     }
 
     @Override
-    public Todo fetchTodoById(Long todoId) throws TodoNotFoundException {
-        Optional<Todo> todo = todoRepository.findById(todoId);
-
-        if (todo.isEmpty()) {
-            throw new TodoNotFoundException("Todo not found");
-        }
-        return todo.get();
+    public Todo fetchTodoById(Long todoId) {
+        return todoRepository.findById(todoId)
+                .orElseThrow(() -> new TodoNotFoundException("Todo with id: " + todoId + " not found"));
     }
 
     @Override
     public Todo saveTodo(Todo todo) {
-        return todoRepository.save(todo);
+        if (getNumberOfIncompleteTasks() >= 10) {
+            throw new IllegalArgumentException("You cannot have more than 10 incomplete tasks,, Complete a task then you can create new ones");
+        } return todoRepository.save(todo);
     }
 
     @Override
     public Todo updateTodoById(Todo updatedTodo, Long todoId) {
-        Todo oldTodo = todoRepository.findById(todoId).get();
+        Todo oldTodo = todoRepository.findById(todoId).orElseThrow(() -> new TodoNotFoundException("Todo with id: " + todoId + " not found"));
 
-        if (Objects.nonNull(updatedTodo.getTask()) &&
-                !updatedTodo.getTask().isEmpty()) {
+        if (Objects.nonNull(updatedTodo.getTask()) && !updatedTodo.getTask().isEmpty()) {
             oldTodo.setTask(updatedTodo.getTask());
         }
 
-        if (Objects.nonNull(updatedTodo.getIsCompleted()) &&
-                !updatedTodo.getIsCompleted().equals(oldTodo.getIsCompleted())) {
+        if (Objects.nonNull(updatedTodo.getIsCompleted()) && !updatedTodo.getIsCompleted().equals(oldTodo.getIsCompleted())) {
             oldTodo.setIsCompleted(updatedTodo.getIsCompleted());
         }
-
         return todoRepository.save(oldTodo);
     }
 
     @Override
     public void deleteTodoById(Long todoId) {
+        if (!todoRepository.existsById(todoId)) {
+            throw new TodoNotFoundException("Todo with id: " + todoId + " not found");
+        }
         todoRepository.deleteById(todoId);
     }
 
@@ -74,8 +72,8 @@ public class TodoServiceImpl implements TodoService {
 
     }
 
-    public Long getNumberOfIncompleteTasks(){
-        return todoRepository.countByIsCompleted(true);
+    public Long getNumberOfIncompleteTasks() {
+        return todoRepository.countByIsCompleted(false);
     }
